requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.12.0
google-cloud-bigquery==3.13.0
